requirements.txt
# ===================================================
# MeetWise Backend v2 — Requirements
# ===================================================

# Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Validation
pydantic>=2.7.0
pydantic-settings>=2.2.0
email-validator>=2.1.0

# LangGraph + LangChain
langgraph>=0.1.19
langchain-core>=0.2.5
langchain-google-genai>=1.0.6

# Google AI SDK (LLM)
google-genai>=0.7.0

# Z3 Theorem Prover
z3-solver>=4.12.0.0

# Google APIs (Workspace + Firebase)
google-auth>=2.29.0
google-api-python-client>=2.129.0
google-cloud-firestore>=2.16.0
httpx>=0.27.0

# Async utilities
anyio>=4.4.0

# Logging
python-json-logger>=2.0.7

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.7
